Three commits land:

1. refactor(R2.1a-2) — body of fleetApplyAction (260 lines) moves into
   internal/fleet.Orchestrator.Run() with 4 methods (resolvePlan,
   resolveVars, installSignalHandler, runSinglePhase). cmd-side
   fleetApplyAction collapses to ~85 lines. ApplyConfig + ApplyRunResult
   structured types let cmd translate ExitCode→cli.Exit at the boundary.
   gocyclo 46→15, cmd/fleet.go -171 LOC. Goroutine topology preserved
   byte-identical.

2. chore(lint) — pragmatic linter slimdown: drop gocritic / unparam /
   dupl / prealloc and staticcheck QF*/ST1005/ST1020/S1008/S1017/S1034
   style categories; keep bug-catching (errcheck / govet / staticcheck
   SA* / ineffassign / unused / unconvert / bodyclose / gosec /
   misspell / cyclop with soft-cap excludes). Fix remaining master
   lint debt (12 issues) inline. `make lint` now clean.

3. chore(docs) — drop preset-examples from docs-generate / docs-check
   after presets/ was retired to ~/projects/mooncake-preset-archive/.

+++ internal/actions/observe_disk/stat_unix.go
@@ -17,7 +17,7 @@ func statPath(path string) (DiskObservation, error) {
 	if err := syscall.Statfs(path, &st); err != nil {
 		return DiskObservation{Path: path}, fmt.Errorf("statfs %s: %w", path, err)
 	}
-	bsize := int64(st.Bsize)
+	bsize := st.Bsize
 	total := int64(st.Blocks) * bsize
 	free := int64(st.Bfree) * bsize
 	avail := int64(st.Bavail) * bsize
